migrate unit_states_test_jboss7 to pytest
